// packages/network/tests/provider/rpc-mapper/methods/eth_sendTransaction/fixture.d.ts
/**
 * Thor solo accounts to use in the tests
 */
declare const THOR_SOLO_ACCOUNTS_ETH_SEND_TRANSACTION_FIXTURE: {
    sender: import("@vechain/sdk-solo-setup").ThorSoloAccount;
    receiver: import("@vechain/sdk-solo-setup").ThorSoloAccount;
};
/**
 * Fixture for a gasPayer private key
 */
declare const gasPayerPrivateKeyFixture: string;
export { THOR_SOLO_ACCOUNTS_ETH_SEND_TRANSACTION_FIXTURE, gasPayerPrivateKeyFixture };
//# sourceMappingURL=fixture.d.ts.map